[dis] Move dispatch into separte directory
